feat: expand metric range from uint16 to uint32
This change allows nylon to have a more fine-grained metric, and enables a workaround for sleeping passive clients

+++ state/routing.go
@@ -68,7 +68,7 @@ type Neighbour struct {
 
 type FD struct {
 	Seqno  uint16
-	Metric uint16
+	Metric uint32
 }
 
 type PubRoute struct {
